Add non-working prevent sleep menu

+++ cantsleep.go
@@ -7,6 +7,7 @@ import (
 	"os/user"
 	"path/filepath"
 	"strconv"
+	"strings"
 	"syscall"
 
 	"github.com/caseymrm/go-assertions"
@@ -82,33 +83,44 @@ func addStartupItem() {
 }
 
 func menuItems() []tray.MenuItem {
-	items := make([]tray.MenuItem, 0)
+	processes := make([]tray.MenuItem, 0)
 	pidAsserts := assertions.GetPIDAssertions()
 	for key := range sleepKeywords {
 		pids := pidAsserts[key]
 		for _, pid := range pids {
-			items = append(items, tray.MenuItem{
-				Text:     fmt.Sprintf("%s (pid %d)", pid.Name, pid.PID),
-				Callback: fmt.Sprintf("%d", pid.PID),
+			processes = append(processes, tray.MenuItem{
+				Text:     pid.Name,
+				Callback: fmt.Sprintf("pid:%d", pid.PID),
 			})
 		}
 	}
-	preAmble := []tray.MenuItem{{Text: "Your laptop can sleep!"}}
-	if len(items) == 1 {
-		preAmble = []tray.MenuItem{{Text: "1 process is keeping your laptop awake:"}}
-	} else if len(items) > 1 {
-		preAmble = []tray.MenuItem{{Text: fmt.Sprintf("%d processes are keeping your laptop awake:", len(items))}}
+	items := []tray.MenuItem{{Text: "Your laptop can sleep!"}}
+	if len(processes) > 0 {
+		if len(processes) == 1 {
+			items = []tray.MenuItem{{Text: "1 process is keeping your laptop awake:"}}
+		} else if len(processes) > 1 {
+			items = []tray.MenuItem{{Text: fmt.Sprintf("%d processes are keeping your laptop awake:", len(processes))}}
+		}
+		items = append(items, processes...)
+		items = append(items, tray.MenuItem{Text: "---"})
 	}
-	if len(items) > 0 {
-		preAmble = append(preAmble, tray.MenuItem{Text: "---"})
-	}
+	items = append(items, tray.MenuItem{
+		Text: "Prevent sleep",
+		Children: []tray.MenuItem{
+			{Text: "5 minutes", Callback: "prevent:5"},
+			{Text: "10 minutes", Callback: "prevent:10"},
+			{Text: "15 minutes", Callback: "prevent:15"},
+			{Text: "30 minutes", Callback: "prevent:30"},
+			{Text: "1 hour", Callback: "prevent:60"},
+			{Text: "2 hours", Callback: "prevent:120"},
+			{Text: "5 hours", Callback: "prevent:300"},
+		},
+	})
 	startupItem := tray.MenuItem{Text: "Run at start up", Callback: "startup"}
 	if runningAtStartup() {
 		startupItem.State = true
 	}
-	items = append(items, startupItem)
-
-	return append(preAmble, items...)
+	return append(items, startupItem)
 }
 
 func menuState() *tray.MenuState {
@@ -140,17 +152,23 @@ func handleClicks(callback chan string) {
 				addStartupItem()
 			}
 		default:
-			pid, _ := strconv.Atoi(clicked)
-			go func() {
-				switch tray.App().Alert("Kill process?", fmt.Sprintf("PID %d", pid), "Kill", "Kill -9", "Cancel") {
-				case 0:
-					fmt.Printf("Killing pid %d\n", pid)
-					syscall.Kill(pid, syscall.SIGTERM)
-				case 1:
-					fmt.Printf("Killing -9 pid %d\n", pid)
-					syscall.Kill(pid, syscall.SIGKILL)
-				}
-			}()
+			if strings.HasPrefix(clicked, "pid:") {
+				pid, _ := strconv.Atoi(clicked[4:])
+				go func() {
+					switch tray.App().Alert("Kill process?", fmt.Sprintf("PID %d", pid), "Kill", "Kill -9", "Cancel") {
+					case 0:
+						fmt.Printf("Killing pid %d\n", pid)
+						syscall.Kill(pid, syscall.SIGTERM)
+					case 1:
+						fmt.Printf("Killing -9 pid %d\n", pid)
+						syscall.Kill(pid, syscall.SIGKILL)
+					}
+				}()
+			}
+			if strings.HasPrefix(clicked, "prevent:") {
+				minutes, _ := strconv.Atoi(clicked[8:])
+				fmt.Printf("prevent %d\n", minutes)
+			}
 		}
 	}
 }
